CV2 Rando Tracker Logic Fix
Fixed issues with the tracker improperly displaying certain checks as green despite not having all the items. Added White Crystal logic to Berkeley Mansion, added laurels and nail or holy water as a requirement for Laruba Mansion, and fixed the Oak Stake checks to require an oak stake in addition to the requirements to enter the mansion. I tested the changes on my own folder and they work.